move inspect into util
