+++ util/mod.ts
@@ -1,6 +1,7 @@
 export * from "./common.ts";
 export * from "./error.ts";
 export * from "./id.ts";
+export * from "./inspect.ts";
 export * from "./map.ts";
 export * from "./memo.ts";
 export * from "./xsync.ts";
